refactor(knowledge): relocate maturityLogic into derive (#209)
