+++ src/architecture/knowledge/derive/maturityLogic.ts
